requirements.txt
aiohttp>=3.10.0
aiofiles>=24.1.0
pika>=1.3.2
pydantic>=2.8.2
pyyaml>=6.0.2
schedule>=1.2.2
click>=8.1.7